#30 - 포스트 삭제시 댓글과 좋아요가 함께 삭제되는 기능 구현
포스트와 댓글, 포스트와 좋아요 모두 단방향으로 매핑되어 있기 때문에, 영속성 전이를 사용하지 않고 각각의 Repository를 호출해 삭제하도록 구현했다.

+++ src/main/java/com/practice/sns/controller/LikeController.java
@@ -24,7 +24,7 @@ public class LikeController {
     }
 
     @GetMapping()
-    public Response<Integer> countLike(@PathVariable Long postId) {
+    public Response<Long> countLike(@PathVariable Long postId) {
         return Response.success(likeService.countLike(postId));
     }
 }
